little test 2

+++ Room Reservation/src/Object/EmployeeInfo.java
@@ -3,9 +3,9 @@ package Object;
 public class EmployeeInfo {
     String name;
     int id;
-    int phone_num;
+    long phone_num;
 
-    public EmployeeInfo(String name,int employee_ID,int phone_num){
+    public EmployeeInfo(String name,int employee_ID,long phone_num){
         this.name = name;
         this.id = employee_ID;
         this.phone_num = phone_num;
@@ -15,7 +15,7 @@ public class EmployeeInfo {
 
     public String getName(){return name;}
 
-    public int getPhoneNum(){return phone_num;}
+    public long getPhoneNum(){return phone_num;}
 
 }
